Refactors recipe conversion work
- Recipe conversion tooling is now broken-out into a subclass of the original
  `RecipeParser`
- This keeps the `RecipeParser file from getting any bigger and will allow some
  future flexibility in maintaining the conversion process
- Test files are broken up accordingly
- Adds `conftest.py` to share testing utilities across multiple unit test files
