+++ tests/parser/test_recipe_parser.py
@@ -1,10 +1,9 @@
 """
 File:           test_recipe_parser.py
-Description:    Unit tests for the recipe parser class and tools
+Description:    Unit tests for the RecipeParser class
 """
 from __future__ import annotations
 
-from pathlib import Path
 from typing import Final
 
 import pytest
@@ -12,11 +11,8 @@ import pytest
 from conda_recipe_manager.parser.enums import SelectorConflictMode
 from conda_recipe_manager.parser.exceptions import JsonPatchValidationException
 from conda_recipe_manager.parser.recipe_parser import RecipeParser
-from conda_recipe_manager.parser.types import MessageCategory
 from conda_recipe_manager.types import JsonType
-
-# Path to supplementary files used in test cases
-TEST_FILES_PATH: Final[str] = "tests/test_aux_files"
+from tests.conftest import TEST_FILES_PATH, load_file, load_recipe
 
 # Long multi-line description string found in the `simple-recipe.yaml` test file
 SIMPLE_DESCRIPTION: Final[
@@ -37,26 +33,6 @@ QUICK_FOX_SUB_PIPE_MINUS: Final[str] = "The quick brown\ntiger\n\njumped over th
 QUICK_FOX_SUB_CARROT: Final[str] = "The quick brown tiger\njumped over the lazy dog\n"
 QUICK_FOX_SUB_CARROT_PLUS: Final[str] = "The quick brown tiger\njumped over the lazy dog\n"
 QUICK_FOX_SUB_CARROT_MINUS: Final[str] = "The quick brown tiger\njumped over the lazy dog"
-
-
-def load_file(file: Path | str) -> str:
-    """
-    Loads a file into a single string
-    :param file: Filename of the file to read
-    :returns: Text from the file
-    """
-    with open(Path(file), "r", encoding="utf-8") as f:
-        return f.read()
-
-
-def load_recipe(file_name: str) -> RecipeParser:
-    """
-    Convenience function that simplifies initializing a recipe parser.
-    :param file_name: File name of the test recipe to load
-    :returns: RecipeParser instance, based on the file
-    """
-    recipe = load_file(f"{TEST_FILES_PATH}/{file_name}")
-    return RecipeParser(recipe)
 
 
 ## Construction and rendering sanity checks ##
@@ -319,90 +295,6 @@ def test_render_to_object_multi_output() -> None:
             },
         ]
     }
-
-
-@pytest.mark.parametrize(
-    "file_base,errors,warnings",
-    [
-        (
-            "simple-recipe.yaml",
-            [],
-            [
-                "A non-list item had a selector at: /requirements/empty_field2",
-                "Required field missing: /about/license_file",
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        (
-            "multi-output.yaml",
-            [],
-            [
-                "Required field missing: /about/summary",
-                "Required field missing: /about/description",
-                "Required field missing: /about/license",
-                "Required field missing: /about/license_file",
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        (
-            "huggingface_hub.yaml",
-            [],
-            [
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        (
-            "types-toml.yaml",
-            [],
-            [
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        # Regression test: Contains a `test` section that caused an empty dictionary to be inserted in the conversion
-        # process, causing an index-out-of-range exception.
-        (
-            "pytest-pep8.yaml",
-            [],
-            [
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        # Regression test: TODO
-        (
-            "google-cloud-cpp.yaml",
-            [],
-            [
-                "A non-list item had a selector at: /outputs/0/script",
-                "A non-list item had a selector at: /outputs/1/script",
-                "A non-list item had a selector at: /outputs/0/script",
-                "A non-list item had a selector at: /outputs/1/script",
-                "Required field missing: /about/description",
-                "Required field missing: /about/license_url",
-            ],
-        ),
-        # TODO Complete: The `curl.yaml` test is far from perfect. It is very much a work in progress.
-        # (
-        #    "curl.yaml",
-        #    [],
-        #    [
-        #        "A non-list item had a selector at: /outputs/0/build/ignore_run_exports",
-        #    ],
-        # ),
-    ],
-)
-def test_render_to_new_recipe_format(file_base: str, errors: list[str], warnings: list[str]) -> None:
-    """
-    Validates rendering a recipe in the new format.
-    :param file_base: Base file name for both the input and the expected out
-    """
-    parser = load_recipe(file_base)
-    result, tbl = parser.render_to_new_recipe_format()
-    assert result == load_file(f"{TEST_FILES_PATH}/new_format_{file_base}")
-    assert tbl.get_messages(MessageCategory.ERROR) == errors
-    assert tbl.get_messages(MessageCategory.WARNING) == warnings
-    # Ensure that the original file was untouched
-    assert not parser.is_modified()
-    assert parser.diff() == ""
 
 
 ## Values ##
